sensors/asset_discovery/requirements.txt
scapy[basic]==2.5.0
python-nmap==0.7.1
kafka-python==2.0.2
python-dotenv==1.0.1